+++ packages/tts-ggml/index.js
@@ -251,6 +251,7 @@ class TTSGgml {
       voiceDir,
       seed,
       nGpuLayers,
+      nCtx,
       threads,
       streamChunkTokens,
       streamFirstChunkTokens,
@@ -355,6 +356,7 @@ class TTSGgml {
     this._voiceDir = voiceDir
     this._seed = seed
     this._nGpuLayers = nGpuLayers
+    this._nCtx = nCtx
     this._threads = threads
     this._streamChunkTokens = streamChunkTokens
     this._streamFirstChunkTokens = streamFirstChunkTokens
@@ -802,6 +804,7 @@ class TTSGgml {
     }
     if (this._seed != null) params.seed = this._seed | 0
     if (this._nGpuLayers != null) params.nGpuLayers = this._nGpuLayers | 0
+    if (this._nCtx != null) params.nCtx = this._nCtx | 0
     if (this._threads != null) params.threads = this._threads | 0
     if (this._streamChunkTokens != null) params.streamChunkTokens = this._streamChunkTokens | 0
     if (this._streamFirstChunkTokens != null) {